+++ index.ts
@@ -31,7 +31,7 @@ export module flounderStyle
     export const styleToString = (style: StyleProperty) => `${style.key.css}: ${style.value ?? "inherit"};`;
     export const styleListToString = (styleList: StyleProperty[], separator: string = " ") =>
         styleList.filter(i => undefined !== i.value).map(i => styleToString(i)).join(separator);
-    export type FlounderType = "tri" | "tetra";
+    export type FlounderType = "trispot" | "tetraspot";
     export type Style = { key: StyleKey; value: StyleValue; };
     export type Color = string;
     export type LayoutAngle = "regular" | "alternative";
@@ -41,21 +41,21 @@ export module flounderStyle
         layoutAngle?: LayoutAngle;
         foregroundColor: Color;
         backgroundColor?: Color; // default is "transparent"
-        spotIntervalSize?: number;
+        intervalSize?: number;
         depth: number; // must be 0.0 <= depth and depth <= 1.0
         blur?: number; // must be 0.0 <= blur
-        maxSpotSize?: number; // must be 1 <= maxSpotSize and maxSpotSize <= (spotIntervalSize *0.5);
+        maxPatternSize?: number; // must be 1 <= maxPatternSize and maxPatternSize <= (intervalSize *0.5);
         reverseRate?: number | "auto"; // must be 0.0 <= depth and depth <= 1.0
         maximumFractionDigits?: number;
     }
-    export const getPatternType = (data: Arguments): FlounderType => data.type ?? "tri";
     export const getLayoutAngle = (data: Arguments): LayoutAngle => data.layoutAngle ?? "regular";
+    export const getPatternType = (data: Arguments): FlounderType => data.type ?? "trispot";
     export const getBackgroundColor = (data: Arguments): Color => data.backgroundColor ?? "transparent";
     export const getBlur = (data: Arguments): number => data.blur ?? config.defaultBlur;
     export const getActualReverseRate = (data: Arguments): number =>
         "number" === typeof data.reverseRate ? data.reverseRate:
-        ("auto" === data.reverseRate && "tri" === getPatternType(data)) ? triPatternHalfRadiusSpotArea:
-        ("auto" === data.reverseRate && "tetra" === getPatternType(data)) ? TetraPatternHalfRadiusSpotArea:
+        ("auto" === data.reverseRate && "trispot" === getPatternType(data)) ? triPatternHalfRadiusSpotArea:
+        ("auto" === data.reverseRate && "tetraspot" === getPatternType(data)) ? TetraPatternHalfRadiusSpotArea:
         999;
     const numberToString = (data: Arguments, value: number) =>
         value.toLocaleString("en-US", { maximumFractionDigits: data.maximumFractionDigits ?? config.defaultMaximumFractionDigits, });
@@ -70,9 +70,9 @@ export module flounderStyle
     {
         switch(getPatternType(data))
         {
-        case "tri":
+        case "trispot":
             return makeTriPatternStyleList(data);
-        case "tetra":
+        case "tetraspot":
             return makeTetraPatternStyleList(data);
         default:
             throw new Error(`Unknown FlounderType: ${data.type}`);
@@ -103,15 +103,15 @@ export module flounderStyle
     const calculateSize = (data: Arguments, halfRadiusSpotArea: number, maxRadiusRate: number) =>
     {
         var radius: number;
-        var spotIntervalSize = data.spotIntervalSize ?? config.defaultSpotIntervalSize;
+        var intervalSize = data.intervalSize ?? config.defaultSpotIntervalSize;
         if (data.depth <= halfRadiusSpotArea)
         {
-            radius = Math.sqrt(data.depth / halfRadiusSpotArea) *(spotIntervalSize *0.5);
+            radius = Math.sqrt(data.depth / halfRadiusSpotArea) *(intervalSize *0.5);
         }
         else
         {
-            const minRadius = spotIntervalSize *0.5;
-            const maxRadius = spotIntervalSize *maxRadiusRate;
+            const minRadius = intervalSize *0.5;
+            const maxRadius = intervalSize *maxRadiusRate;
             const MaxRadiusWidth = maxRadius -minRadius;
             const minAreaRate = 1.0 -Math.sqrt(1.0 -halfRadiusSpotArea);
             const maxAreaRate = 1.0;
@@ -120,12 +120,12 @@ export module flounderStyle
             const areaRateWidth = areaRate -minAreaRate;
             radius = minRadius +(MaxRadiusWidth *Math.pow(areaRateWidth / maxAreaRateWidth, 2));
         }
-        if (undefined !== data.maxSpotSize && data.maxSpotSize < radius)
+        if (undefined !== data.maxPatternSize && data.maxPatternSize < radius)
         {
-            spotIntervalSize = spotIntervalSize *data.maxSpotSize /radius;
-            radius = data.maxSpotSize;
+            intervalSize = intervalSize *data.maxPatternSize /radius;
+            radius = data.maxPatternSize;
         }
-        return { radius, spotIntervalSize };
+        return { radius, intervalSize };
     };
     export const reverseArguments = (data: Arguments): Arguments =>
     {
@@ -158,7 +158,7 @@ export module flounderStyle
         }
         else
         {
-            const { radius, spotIntervalSize } = calculateSize(data, triPatternHalfRadiusSpotArea, 1.0 /root3);
+            const { radius, intervalSize } = calculateSize(data, triPatternHalfRadiusSpotArea, 1.0 /root3);
             const radialGradient = makeRadialGradientString(data, radius);
             const backgroundColor: StyleValue = getBackgroundColor(data);
             const backgroundImage: StyleValue = Array.from({ length: 4 }).map(_ => radialGradient).join(", ");
@@ -169,16 +169,16 @@ export module flounderStyle
                 ({
                     backgroundColor,
                     backgroundImage,
-                    backgroundSize: `${numberToString(data, spotIntervalSize *2.0)}px ${numberToString(data, spotIntervalSize *root3)}px`,
-                    backgroundPosition: `0px 0px, ${numberToString(data, spotIntervalSize)}px 0px, ${numberToString(data, spotIntervalSize *0.5)}px ${numberToString(data, spotIntervalSize *root3 *0.5)}px, ${numberToString(data, spotIntervalSize *1.5)}px ${numberToString(data, spotIntervalSize * root3 * 0.5)}px`
+                    backgroundSize: `${numberToString(data, intervalSize *2.0)}px ${numberToString(data, intervalSize *root3)}px`,
+                    backgroundPosition: `0px 0px, ${numberToString(data, intervalSize)}px 0px, ${numberToString(data, intervalSize *0.5)}px ${numberToString(data, intervalSize *root3 *0.5)}px, ${numberToString(data, intervalSize *1.5)}px ${numberToString(data, intervalSize * root3 * 0.5)}px`
                 });
             case "alternative": // vertical
                 return makeResult
                 ({
                     backgroundColor,
                     backgroundImage,
-                    backgroundSize: ` ${numberToString(data, spotIntervalSize *root3)}px ${numberToString(data, spotIntervalSize *2.0)}px`,
-                    backgroundPosition: `0px 0px, 0px ${numberToString(data, spotIntervalSize)}px, ${numberToString(data, spotIntervalSize *root3 *0.5)}px ${numberToString(data, spotIntervalSize *0.5)}px, ${numberToString(data, spotIntervalSize *root3 *0.5)}px ${numberToString(data, spotIntervalSize *1.5)}px`
+                    backgroundSize: ` ${numberToString(data, intervalSize *root3)}px ${numberToString(data, intervalSize *2.0)}px`,
+                    backgroundPosition: `0px 0px, 0px ${numberToString(data, intervalSize)}px, ${numberToString(data, intervalSize *root3 *0.5)}px ${numberToString(data, intervalSize *0.5)}px, ${numberToString(data, intervalSize *root3 *0.5)}px ${numberToString(data, intervalSize *1.5)}px`
                 });
             default:
                 throw new Error(`Unknown LayoutAngle: ${data.layoutAngle}`);
@@ -207,7 +207,7 @@ export module flounderStyle
         }
         else
         {
-            const { radius, spotIntervalSize } = calculateSize(data, TetraPatternHalfRadiusSpotArea, 0.5 *root2);
+            const { radius, intervalSize } = calculateSize(data, TetraPatternHalfRadiusSpotArea, 0.5 *root2);
             const radialGradient = makeRadialGradientString(data, radius);
             const backgroundColor: StyleValue = getBackgroundColor(data);
             switch(getLayoutAngle(data))
@@ -217,15 +217,15 @@ export module flounderStyle
                 ({
                     backgroundColor,
                     backgroundImage: radialGradient,
-                    backgroundSize: `${numberToString(data, spotIntervalSize)}px ${numberToString(data, spotIntervalSize)}px`,
+                    backgroundSize: `${numberToString(data, intervalSize)}px ${numberToString(data, intervalSize)}px`,
                 });
             case "alternative": // slant
                 return makeResult
                 ({
                     backgroundColor,
                     backgroundImage: Array.from({ length: 2 }).map(_ => radialGradient).join(", "),
-                    backgroundSize: `${numberToString(data, (spotIntervalSize *2.0) /root2)}px ${numberToString(data, (spotIntervalSize *2.0) /root2)}px`,
-                    backgroundPosition: `0px 0px, ${numberToString(data, spotIntervalSize /root2)}px ${numberToString(data, spotIntervalSize /root2)}px`
+                    backgroundSize: `${numberToString(data, (intervalSize *2.0) /root2)}px ${numberToString(data, (intervalSize *2.0) /root2)}px`,
+                    backgroundPosition: `0px 0px, ${numberToString(data, intervalSize /root2)}px ${numberToString(data, intervalSize /root2)}px`
                 });
             default:
                 throw new Error(`Unknown LayoutAngle: ${data.layoutAngle}`);